fastapi-python-backend/pyproject.toml
[build-system]
requires = ["setuptools>=69.0"]
build-backend = "setuptools.build_meta"

[project]
name = "fastapi-python-backend-template"
version = "1.0.0"
description = "Production-focused FastAPI + Python backend template"
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
  "fastapi>=0.116.0,<1.0.0",
  "prometheus-client>=0.23.1,<1.0.0",
  "pydantic-settings>=2.11.0,<3.0.0",
  "python-json-logger>=2.0.7,<3.0.0",
  "uvicorn[standard]>=0.35.0,<1.0.0"
]

[project.optional-dependencies]
dev = [
  "httpx>=0.28.1,<1.0.0",
  "mypy>=1.17.1,<2.0.0",
  "pytest>=8.4.1,<9.0.0",
  "pytest-asyncio>=1.1.0,<2.0.0",
  "ruff>=0.12.8,<1.0.0"
]

[tool.setuptools]
packages = ["src"]

[tool.pytest.ini_options]
pythonpath = ["."]
testpaths = ["tests"]

[tool.ruff]
line-length = 100
target-version = "py311"

[tool.ruff.lint]
select = ["E", "F", "I", "B", "UP"]

[tool.ruff.format]
quote-style = "single"

[tool.mypy]
python_version = "3.11"
warn_unused_ignores = true
warn_redundant_casts = true
strict_optional = true
check_untyped_defs = true
ignore_missing_imports = true

